do-release-tags: Add remote pull option Allows users to automate the pull and promotion of content from external repositories.  This adds two new optional YAML key value pair 'remote' and 'sourceref'.
The remote YAML key can also be specified as a command line argument (--remote)

sourceref is the refernce in the remote repositoy to pull.  Because we are explicitly specifying the commit to pull it doesn't actually doesn't matter what value you use. You will probably want to set it to the name of the reference used upstream because it will show up when you run `ostree refs` on your destination repository. If sourceref isn't specified it defaults to baseref/release.